+++ packages/mini_program_tooling/test/miniprogram_cli_test.dart
@@ -801,7 +801,7 @@ dependencies:
       );
       expect(
         await File(p.join(projectRoot, 'pubspec.yaml')).readAsString(),
-        contains('mini_program_sdk: ^0.1.1'),
+        contains('mini_program_sdk: ^0.1.2'),
       );
     });